13/7 explore more n some small issues

+++ lib/model/responseModek/for_program_page_response_model.dart
@@ -71,7 +71,7 @@ class ForProgramPageResponseModel {
   String? instituteAbout;
   String? instAddress;
   String? programDuration;
-  int? applicationFee;
+  double? applicationFee;
   String? instTile;
   int? instCountryId;
   String? specializationName;
